From MUI TextField to MUI DatePicker
- I set the state to the format I need to have in my db
If I didn't do that then it'll show the unnecessary Timezone with it. Which messes up my result.
If I choose 18/11/2023 for example the TimeZone picks that up and I get 17/11/2023 as a result

I just need this format('YYYY-MM-DD') to be saved in the database and not this

startdate: M $D: 12 $H0 $L: "nl-be" $M: 10 $W: 0 $d: Sat Nov 18 2023 00:00:00 GMT+0100 (Mid-European TimeZone) $isDayjsObject: true $m: 0 $ms: 0 $s: 0 $u: undefined $x: {} $y: 2023