addedTeamIdToTeamplan migration
If you want to create a teamplan you also have to select which team you want to dedicate this teamplan to. In the NurseShiftPage if you add another NurseShift you don't have the option anymore to select which team you're going to put. The specific teamplan is only specified to a specific team.

Added the teamplanId to the ExcelController